+++ js/consts.js
@@ -1,0 +1,38 @@
+const NAMES = [
+  'Кирилл',
+  'Софья',
+  'Юлия',
+  'Арина',
+  'Виктор',
+  'Светлана',
+  'Даниил',
+  'Вероника',
+  'Дмитрий',
+  'Алёна',
+];
+
+const USERS_COUNT = 25;
+const COMMENTS_COUNT_MAX = 6;
+const COMMENTS_COUNT_MIN = 2;
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const DESCRIPTION = [
+  'Это я в горах, загораю',
+  'Кому привезти открытку с отпуска?',
+  'Жизнь прекрасна!',
+  'Сегодня думал о смысле жизни, ничего придумать не удалось.',
+  'Обычно здесь что-то пишут, но я пока не придумал',
+  'Моя любимая фотка',
+];
+
+export { NAMES, USERS_COUNT, COMMENTS_COUNT, COMMENTS, DESCRIPTION, MIN_LIKES, MAX_LIKES };
